Refactor: merge duplicate controller directories
